convert to tsx
