+++ tgui/packages/tgui/interfaces/Secrets/types.ts
@@ -1,0 +1,3 @@
+import type { BooleanLike } from 'tgui-core/react';
+
+export type Data = { is_debugger: BooleanLike; is_funmin: BooleanLike };
